Refactor LoadingSpinner import paths and restore LoadingSpinner component
